refactor(docs-page): to TS; split up file
