+++ packages/docs-page/server/index.ts
@@ -1,0 +1,11 @@
+// We currently export most utilities individually,
+// since we have cases such as Packer remote plugin docs
+// where we want to re-use these utilities to build
+// getStaticPaths and getStaticProps functions that
+// fall outside the use case of local-only content
+export { generateStaticPaths } from './generate-static-paths'
+export { generateStaticProps } from './generate-static-props'
+export { getNodeFromPath } from './get-node-from-path'
+export { getPathsFromNavData } from './get-paths-from-nav-data'
+export { validateNavData } from './validate-nav-data'
+export { validateFilePaths } from '@hashicorp/react-docs-sidenav/utils/validate-file-paths'
